# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface CreateVideoBody {
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minimum 1 */
  sizeBytes: number;
  /** @minLength 1 */
  contentType: string;
}

export interface Video {
  id: string;
  fileName: string;
  objectPath: string;
  sizeBytes: number;
  durationSeconds?: number | null;
  contentType?: string;
  /** pending | processing | done | failed */
  status: string;
  createdAt: string;
}

export interface Job {
  id: string;
  videoId: string;
  /** queued | processing | done | failed */
  status: string;
  progressPct: number;
  /** Human-readable label of the current pipeline stage */
  stage: string;
  errorMessage?: string | null;
  startedAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface Clip {
  id: string;
  videoId: string;
  rank: number;
  startSec: number;
  endSec: number;
  durationSec: number;
  viralScore: number;
  hookScore: number;
  energyScore: number;
  sentimentScore: number;
  keywordDensity: number;
  speechRate: number;
  caption: string;
  hookLine: string;
  transcriptText: string;
  rationale: string;
}

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
}

export interface Transcript {
  id: string;
  videoId: string;
  fullText: string;
  language: string;
  segments: TranscriptSegment[];
}

export interface VideoWithJob {
  video: Video;
  job: Job;
}

export interface VideoDetail {
  video: Video;
  job: Job;
  clips: Clip[];
}

export interface Stats {
  totalVideos: number;
  totalClips: number;
  completedVideos: number;
  processingVideos: number;
  failedVideos: number;
  averageViralScore: number;
  topClips: Clip[];
}