# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Request a presigned URL for file upload
 */

export const RequestUploadUrlBody = zod.object({
  name: zod.string().min(1),
  size: zod.number().min(1),
  contentType: zod.string().min(1),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
  metadata: zod
    .object({
      name: zod.string().min(1),
      size: zod.number().min(1),
      contentType: zod.string().min(1),
    })
    .optional(),
});

/**
 * @summary Serve a public asset
 */
export const GetPublicObjectParams = zod.object({
  filePath: zod.coerce.string(),
});

/**
 * @summary Serve an object entity
 */
export const GetStorageObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});

/**
 * @summary List all uploaded videos
 */
export const ListVideosResponseItem = zod.object({
  id: zod.string(),
  fileName: zod.string(),
  objectPath: zod.string(),
  sizeBytes: zod.number(),
  durationSeconds: zod.number().nullish(),
  contentType: zod.string().optional(),
  status: zod.string().describe("pending | processing | done | failed"),
  createdAt: zod.coerce.date(),
});
export const ListVideosResponse = zod.array(ListVideosResponseItem);

/**
 * @summary Register a newly uploaded video and start processing
 */

export const CreateVideoBody = zod.object({
  fileName: zod.string().min(1),
  objectPath: zod.string().min(1),
  sizeBytes: zod.number().min(1),
  contentType: zod.string().min(1),
});

/**
 * @summary Get a single video with its job and clips
 */
export const GetVideoParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVideoResponse = zod.object({
  video: zod.object({
    id: zod.string(),
    fileName: zod.string(),
    objectPath: zod.string(),
    sizeBytes: zod.number(),
    durationSeconds: zod.number().nullish(),
    contentType: zod.string().optional(),
    status: zod.string().describe("pending | processing | done | failed"),
    createdAt: zod.coerce.date(),
  }),
  job: zod.object({
    id: zod.string(),
    videoId: zod.string(),
    status: zod.string().describe("queued | processing | done | failed"),
    progressPct: zod.number(),
    stage: zod
      .string()
      .describe("Human-readable label of the current pipeline stage"),
    errorMessage: zod.string().nullish(),
    startedAt: zod.coerce.date().nullish(),
    completedAt: zod.coerce.date().nullish(),
    createdAt: zod.coerce.date(),
  }),
  clips: zod.array(
    zod.object({
      id: zod.string(),
      videoId: zod.string(),
      rank: zod.number(),
      startSec: zod.number(),
      endSec: zod.number(),
      durationSec: zod.number(),
      viralScore: zod.number(),
      hookScore: zod.number(),
      energyScore: zod.number(),
      sentimentScore: zod.number(),
      keywordDensity: zod.number(),
      speechRate: zod.number(),
      caption: zod.string(),
      hookLine: zod.string(),
      transcriptText: zod.string(),
      rationale: zod.string(),
    }),
  ),
});

/**
 * @summary Delete a video and all of its clips
 */
export const DeleteVideoParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get the latest processing job for a video (status, progress, stage)
 */
export const GetVideoJobParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVideoJobResponse = zod.object({
  id: zod.string(),
  videoId: zod.string(),
  status: zod.string().describe("queued | processing | done | failed"),
  progressPct: zod.number(),
  stage: zod
    .string()
    .describe("Human-readable label of the current pipeline stage"),
  errorMessage: zod.string().nullish(),
  startedAt: zod.coerce.date().nullish(),
  completedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all viral clips found for a video, ranked by viral score
 */
export const ListVideoClipsParams = zod.object({
  id: zod.coerce.string(),
});

export const ListVideoClipsResponseItem = zod.object({
  id: zod.string(),
  videoId: zod.string(),
  rank: zod.number(),
  startSec: zod.number(),
  endSec: zod.number(),
  durationSec: zod.number(),
  viralScore: zod.number(),
  hookScore: zod.number(),
  energyScore: zod.number(),
  sentimentScore: zod.number(),
  keywordDensity: zod.number(),
  speechRate: zod.number(),
  caption: zod.string(),
  hookLine: zod.string(),
  transcriptText: zod.string(),
  rationale: zod.string(),
});
export const ListVideoClipsResponse = zod.array(ListVideoClipsResponseItem);

/**
 * @summary Get the full transcript and segment timestamps for a video
 */
export const GetVideoTranscriptParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVideoTranscriptResponse = zod.object({
  id: zod.string(),
  videoId: zod.string(),
  fullText: zod.string(),
  language: zod.string(),
  segments: zod.array(
    zod.object({
      start: zod.number(),
      end: zod.number(),
      text: zod.string(),
    }),
  ),
});

/**
 * @summary Re-run the viral analysis pipeline for an existing video
 */
export const ReprocessVideoParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get dashboard summary statistics
 */
export const GetStatsResponse = zod.object({
  totalVideos: zod.number(),
  totalClips: zod.number(),
  completedVideos: zod.number(),
  processingVideos: zod.number(),
  failedVideos: zod.number(),
  averageViralScore: zod.number(),
  topClips: zod.array(
    zod.object({
      id: zod.string(),
      videoId: zod.string(),
      rank: zod.number(),
      startSec: zod.number(),
      endSec: zod.number(),
      durationSec: zod.number(),
      viralScore: zod.number(),
      hookScore: zod.number(),
      energyScore: zod.number(),
      sentimentScore: zod.number(),
      keywordDensity: zod.number(),
      speechRate: zod.number(),
      caption: zod.string(),
      hookLine: zod.string(),
      transcriptText: zod.string(),
      rationale: zod.string(),
    }),
  ),
});