# lib/api-zod/src/generated/types/clip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Clip {
  id: string;
  videoId: string;
  rank: number;
  startSec: number;
  endSec: number;
  durationSec: number;
  viralScore: number;
  hookScore: number;
  energyScore: number;
  sentimentScore: number;
  keywordDensity: number;
  speechRate: number;
  caption: string;
  hookLine: string;
  transcriptText: string;
  rationale: string;
}