# lib/api-zod/src/generated/types/createVideoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoBody {
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minimum 1 */
  sizeBytes: number;
  /** @minLength 1 */
  contentType: string;
}