# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: string;
  videoId: string;
  /** queued | processing | done | failed */
  status: string;
  progressPct: number;
  /** Human-readable label of the current pipeline stage */
  stage: string;
  errorMessage?: string | null;
  startedAt?: Date | null;
  completedAt?: Date | null;
  createdAt: Date;
}