# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Clip } from "./clip";

export interface Stats {
  totalVideos: number;
  totalClips: number;
  completedVideos: number;
  processingVideos: number;
  failedVideos: number;
  averageViralScore: number;
  topClips: Clip[];
}