# lib/api-zod/src/generated/types/transcriptSegment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
}