# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: string;
  fileName: string;
  objectPath: string;
  sizeBytes: number;
  durationSeconds?: number | null;
  contentType?: string;
  /** pending | processing | done | failed */
  status: string;
  createdAt: Date;
}