# lib/api-zod/src/generated/types/videoDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Clip } from "./clip";
import type { Job } from "./job";
import type { Video } from "./video";

export interface VideoDetail {
  video: Video;
  job: Job;
  clips: Clip[];
}