# lib/api-zod/src/generated/types/videoWithJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auto Viral Clip Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";
import type { Video } from "./video";

export interface VideoWithJob {
  video: Video;
  job: Job;
}